app_shortener/requirements.txt
alembic==1.9.2
anyio==3.6.2
click==8.1.3
decorator==5.1.1
fastapi==0.89.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.5
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.18.1
websockets==10.4

